test fixture version of shim class
